refactor: update retriever name
